Move install wizard enums
