+++ main/src/cgeo/geocaching/wizard/WizardStep.java
@@ -1,0 +1,10 @@
+package cgeo.geocaching.wizard;
+
+public enum WizardStep {
+    WIZARD_START,
+    WIZARD_PERMISSIONS, WIZARD_PERMISSIONS_STORAGE, WIZARD_PERMISSIONS_LOCATION,
+    WIZARD_PERMISSIONS_BASEFOLDER, WIZARD_PERMISSIONS_MAPFOLDER, WIZARD_PERMISSIONS_MAPTHEMEFOLDER, WIZARD_PERMISSIONS_GPXFOLDER, WIZARD_PERMISSIONS_BROUTERTILESFOLDER,
+    WIZARD_PLATFORMS,
+    WIZARD_ADVANCED,
+    WIZARD_END
+}
